Change escrow hooks file name to make more sense and update some inner logic to use the position and historical positions queries instead of the escrow query hook previously used
